requirements.txt
flask==2.0.2
werkzeug==2.0.2
google-cloud-firestore==2.3.2
mysql-connector-python==8.0.27
flask_cors==3.0.10
flask_mail
tensorflow
scikit-surprise
pandas
joblib
scikit-learn